pyproject.toml
[tool.poetry]
name = "rag"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [
    { include = "app" },
]

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "^0.23.2"
langserve = {extras = ["server"], version = ">=0.0.30"}
pydantic = ">=2.7.4,<3.0.0"
langchain-community = "^0.3.19"
tqdm = "^4.67.1"
unstructured = {extras = ["all-docs"], version = "^0.16.25"}
langchain-experimental = "^0.3.4"
langchain-openai = "^0.3.8"
python-dotenv = "^1.0.1"
tiktoken = "^0.9.0"
psycopg = "^3.2.5"
pgvector = "^0.3.6"
langchain-postgres = "^0.0.13"
fastapi = "^0.115.11"
pymupdf = "^1.25.3"
aiofiles = "^24.1.0"


[tool.poetry.group.dev.dependencies]
langchain-cli = ">=0.0.15"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
